add rotas lateral
