+++ src/Router.tsx
@@ -10,22 +10,32 @@ import { SignUp } from "./pages/SignUp";
 import { Perfil } from "./pages/Perfil";
 import CadastroMotorista from "./pages/SignUpMotoristas"; 
 import CadastroVeiculo from "./pages/SignUpVeiculos"; 
+import { BarraLateralLayout } from "./layouts/BarraLateralFrotaLayout";
+import { BarraLateralMotoristaLayout } from "./layouts/BarraLateralLayoutMotoristas";
 
 export function Router() {
   return (
     <Routes>
         <Route path="/" element={<DefaultLayout />}>
             <Route path="/" element={<Home />} />
-            <Route path="/frotas" element={<Frotas />} />
-            <Route path="/motoristas" element={<Motoristas />} />
-            <Route path="/cadastro-veiculo" element={<CadastroVeiculo />} />
-            <Route path="/cadastro-motorista" element={<CadastroMotorista />} /> {/* ✅ atualizado */}
+
+
         </Route>
         <Route path="/" element={<SignLayout />}>
             <Route path="/sign" element={<Sign />} />
             <Route path="/signup" element={<SignUp />} />
             <Route path="/perfil" element={<Perfil />} />
         </Route>
+        <Route path="/" element={<BarraLateralLayout />}>
+            <Route path="/frotas" element={<Frotas />} />
+            <Route path="/cadastro-veiculo" element={<CadastroVeiculo />} />
+        </Route>
+        <Route path="/" element={<BarraLateralMotoristaLayout />}>
+            <Route path="/motoristas" element={<Motoristas />} />
+            <Route path="/cadastro-motorista" element={<CadastroMotorista />} />
+        </Route>
+
+        <Route path="*" element={<h1>Página não encontrada</h1>} />
     </Routes>
   );
 }
