+++ src/layouts/BarraLateralFrotaLayout/styles.ts
@@ -4,5 +4,6 @@ export const LayoutContainer = styled.div`
     background-color:rgb(255, 255, 255);
     display: flex;
     min-height: 100vh;
+    
 
     `;
