+++ src/layouts/DefaultLayout/styles.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-export const LayoutContainer = styled.div
-    `background-color:rgb(255, 255, 255);
+export const LayoutContainer = styled.div`
+    background-color:rgb(255, 255, 255);
     display: flex;
     flex-direction: column;
     min-height: 100vh;
