+++ src/pages/SignUpMotoristas/styles.ts
@@ -15,35 +15,7 @@ export const MotoristasContainer = styled.div`
   background: #0d1117;
   color: #E8EBED;
   min-height: 100vh;
-`;
-
-export const BarraLateral = styled.div`
-  width: 250px;
-  background: #161b22;
-  padding: 30px 20px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-
-  ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-
-    li {
-      margin-bottom: 20px;
-      cursor: pointer;
-      font-weight: 600;
-      color: #7A8A96;
-      font-size: 16px;
-      transition: color 0.3s, transform 0.3s;
-
-      &:hover {
-        color: #DE562C;
-        transform: translateX(5px);
-      }
-    }
-  }
+  width: 100%;
 `;
 
 export const PainelMotoristas = styled.div`
