--- src/pages/SignUpVeiculos/types.ts
@@ -1,6 +1,0 @@
-export interface Veiculo {
-  placa: string;
-  codigoVeiculo: string;
-  cor: string;
-  categoria: string;
-}
